Added Yetsira dimensions tree diagram.
This is a diagram of my own interpretation of Yetsira mapped
onto a Golden-Dawnesque Tree of Life.  It was interesting to
note that the inner cube axes (mother letters) map onto this
tree via reflections of their standard paths.  This is drawn
into the tree.

The numbers on the sephirot indicate the order of the sephirot
described in Yetsira. It has clearly been mangled on its way
to the tree form, but I think that's OK.