Added Wine-Women-Song Virtue tree
This is from Mandrake the Bard (Ultima VI), who had an alternate
set of principles and virtues.  These are mapped to the tree the
exact same way that the typical Ultima virtues are in the other
diagram.  The only one that seems far out of place is "Laziness"
for Geburah, but one can smooth over the difference by thinking
of Laziness as a form of holding back and restricting oneself
from activities.