Starting to make the SVGs consistent on Safari
The main thing I hit was: in safari, the text-anchor and
dominant-baseline are not inherited from the parent <g>
group... so to get the same effect without repeating the
attributes, I used CSS instead, and put the groups in a
CSS class.

There are still more diagrams to correct, to come in a future
commit.